+++ Trarizon.Library/Collections/Helpers/ListQueries/Helpers/.cs
@@ -1,6 +1,5 @@
 ﻿using System.Diagnostics.CodeAnalysis;
 using System.Runtime.CompilerServices;
-using Trarizon.Library.Collections.Helpers.ListQueries.Helpers;
 
 namespace Trarizon.Library.Collections.Helpers;
 [SuppressMessage("Style", "IDE0301")] // Use explicit on Empty<T>(), collection will alloc new List<T> in some conditions