+++ Trarizon.Library/Collections/Helpers/ListQueries/Select.cs
@@ -4,7 +4,6 @@
 
 using System.Diagnostics;
 using Trarizon.Library.Collections.Helpers.Utilities.Queriers;
-using Trarizon.Library.Collections.Helpers.ListQueries.Helpers;
 using Trarizon.Library.Collections.StackAlloc;
 
 namespace Trarizon.Library.Collections.Helpers;